testmailchimp/requirements.txt
django
djangorestframework
mailchimp3